Wallet/requirements.txt
matplotlib
django-mptt